+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Changed position of menu items in Admin to be more logical
 - Split Add and Import collection pages, reflected in navigation
 - Swapped manifests / collections order in admin
+- Moved `<DocumentPreview />` from capture models to this repo
 
 ### Removed